Rename the existing commands package
