+++ src/main/java/com/craftyn/casinoslots/CasinoSlots.java
@@ -19,11 +19,11 @@ import com.craftyn.casinoslots.classes.SlotMachine;
 import com.craftyn.casinoslots.classes.SlotMachineMessages;
 import com.craftyn.casinoslots.classes.SlotMachineOwner;
 import com.craftyn.casinoslots.classes.SlotType;
-import com.craftyn.casinoslots.command.AnCommandExecutor;
 import com.craftyn.casinoslots.enums.Settings;
 import com.craftyn.casinoslots.listeners.BlockListener;
 import com.craftyn.casinoslots.listeners.EntityListener;
 import com.craftyn.casinoslots.listeners.PlayerListener;
+import com.craftyn.casinoslots.oldcommands.AnCommandExecutor;
 import com.craftyn.casinoslots.slot.SlotManager;
 import com.craftyn.casinoslots.slot.TypeManager;
 import com.craftyn.casinoslots.util.ConfigData;
