+++ src/main/java/com/craftyn/casinoslots/oldcommands/AnCommandExecutor.java
@@ -1,4 +1,4 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
