+++ src/main/java/com/craftyn/casinoslots/oldcommands/CasinoAdd.java
@@ -1,6 +1,5 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
-import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
 import com.craftyn.casinoslots.CasinoSlots;
@@ -8,20 +7,19 @@ import com.craftyn.casinoslots.classes.OldSlotMachine;
 import com.craftyn.casinoslots.classes.SlotType;
 import com.craftyn.casinoslots.util.PermissionUtil;
 
-public class CasinoAddManaged extends AnCommand {
+public class CasinoAdd extends AnCommand {
 
-    /**
-     * Instantiates a new managed CasinoSlot.
-     *
-     * @param plugin  The plugin.
-     * @param args    The args sent with the command.
-     * @param player  The player doing the command.
-     */
-    public CasinoAddManaged (CasinoSlots plugin, String[] args, Player player) {
+    // Command for adding unmanaged slot machine
+    public CasinoAdd(CasinoSlots plugin, String[] args, Player player) {
         super(plugin, args, player);
     }
 
     public boolean process() {
+        // Permissions
+        if(!PermissionUtil.canCreate(player)) {
+            noPermission();
+            return true;
+        }
 
         //Check for simple player things before they try to add a slot
         if(plugin.useTowny) {
@@ -41,58 +39,57 @@ public class CasinoAddManaged extends AnCommand {
         }
 
         // Valid command format
-        if(args.length == 3) {
+        if(args.length >= 2 && args.length <= 3) {
 
             // Slot does not exist
             if(!plugin.getSlotManager().isSlot(args[1])) {
                 SlotType type;
 
-                // Valid type
-                if(plugin.getTypeManager().isValidType(args[2])) {
-                    String typeName = args[2];
-
-                    // Has type permission
-                    if(PermissionUtil.canCreateManagedType(player, typeName)) {
-                        type = plugin.getTypeManager().getType(typeName);
+                // Check to see if the type is valid
+                if(args.length < 3) {
+                    type = plugin.getTypeManager().getType("default");
+                } else if(plugin.getTypeManager().isValidType(args[2])) {
+                    if(PermissionUtil.canCreate(player, args[2])) {
+                        type = plugin.getTypeManager().getType(args[2]);
                     } else {
-                        plugin.sendMessage(player, ChatColor.RED + "You do not have permission to create a managed slot.");
+                        sendMessage("You do not have permission to create a machine with the type: " + args[2]);
                         return true;
                     }
-                }
-
-                // Invalid type
-                else {
-                    plugin.sendMessage(player, "Invalid type " + args[2]);
+                } else {
+                    // Invalid type
+                    sendMessage("Invalid type " + args[2]);
                     return true;
                 }
-
+                
                 // Creation cost
                 Double createCost = type.getCreateCost();
                 if(plugin.getEconomy().has(player, createCost)) {
                     plugin.getEconomy().withdrawPlayer(player, createCost);
                 } else {
-                    sendMessage("You can't afford to create this slot machine. Cost: " + createCost);
+                    sendMessage(player.getName());
+                    sendMessage("You do not have enough to afford to create this slot machine. The cost is: " + createCost);
                     return true;
                 }
 
                 //Good to start punching the blocks to create the slot.
-                OldSlotMachine slot = new OldSlotMachine(args[1], type, player.getUniqueId(), player.getName(), player.getWorld().getName(), true, false, 0, 0);
+                OldSlotMachine slot = new OldSlotMachine(args[1], type, player.getUniqueId(), player.getName(), player.getWorld().getName(), false, false, 0, 0);
                 plugin.getSlotManager().toggleCreatingSlots(player.getName(), slot);
-                plugin.sendMessage(player, "Punch a block to serve as the base for this slot machine.");
+                sendMessage("Punch a block to serve as the base for this slot machine.");
             }
 
             // Slot exists
             else {
-                plugin.sendMessage(player, "Slot machine " + args[1] +" already registered.");
+                sendMessage("A slot machine by the name of \"" + args[1] + "\" is already registered.");
             }
         }
 
         // incorrect command format
         else {
-            plugin.sendMessage(player, "Usage:");
-            plugin.sendMessage(player, "/casino addmanaged <name> <type>");
+            sendMessage("Usage:");
+            sendMessage("/casino add <name> (<type>)");
         }
         return true;
 
     }
+
 }
