+++ src/main/java/com/craftyn/casinoslots/oldcommands/CasinoAddItem.java
@@ -1,4 +1,4 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
