+++ src/main/java/com/craftyn/casinoslots/oldcommands/CasinoSetOwner.java
@@ -1,4 +1,4 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
 import org.bukkit.entity.Player;
