+++ src/main/java/com/craftyn/casinoslots/oldcommands/CasinoStats.java
@@ -1,4 +1,4 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
 import java.util.Collection;
