+++ src/main/java/com/craftyn/casinoslots/oldcommands/CasinoType.java
@@ -1,4 +1,4 @@
-package com.craftyn.casinoslots.command;
+package com.craftyn.casinoslots.oldcommands;
 
 import org.bukkit.entity.Player;
